error handling,more test cases,fixed the database persisting issue.updated Readme

+++ src/main/java/com/task/management/repositories/TaskRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 import com.task.management.entity.Task;
 
 @Repository
-public interface TaskRepository extends CrudRepository<Task, Long>, QueryByExampleExecutor<Task> {
+public interface TaskRepository extends CrudRepository<Task, Integer>, QueryByExampleExecutor<Task> {
 }
