+++ src/main/java/com/task/management/service/TaskServiceImpl.java
@@ -25,7 +25,7 @@ public class TaskServiceImpl implements TaskService {
   }
 
   @Override
-  public Optional<Task> getTaskById(Long id) {
+  public Optional<Task> getTaskById(int id) {
     return taskRepository.findById(id);
   }
 
@@ -41,7 +41,7 @@ public class TaskServiceImpl implements TaskService {
   }
 
   @Override
-  public void deleteTask(Long id) {
+  public void deleteTask(int id) {
     taskRepository.deleteById(id);
   }
